+++ packages/plugins/eslint-plugin-react-x/src/rules/no-complex-conditional-rendering.ts
@@ -11,7 +11,6 @@ export type MessageID = CamelCase<typeof RULE_NAME>;
 export default createRule<[], MessageID>({
   meta: {
     type: "problem",
-    deprecated: true,
     docs: {
       description: "disallow complex conditional rendering",
     },